pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-sql-agent"
version = "1.0.0"
description = "AI agent for generating Redshift SQL queries"
authors = [{name = "Author", email = "author@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ollama==0.3.3",
    "python-dotenv==1.0.0"
]
keywords = ["ai", "sql", "redshift", "agent"]

[project.scripts]
ai-sql-agent = "ai_agent_py.main:main"